Fix teardown deadlock
Don't wait on the readers from the session teardown anymore, because if
the only thread that can complete reading the buffer is the same waiting
on this condition, it deadlocks.

If the session daemon needs to wait on consumers, it can always do that
in user-space through some other mechanism.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>