Fix: tracepoint listing misses last loaded probe
The seqfile iteration listing tracepoints operates on the probe list
directly, without going through the lttng_get_probe_list_head()
accessor which ensures that the lazy list of probes is moved to the
actual list of probes.

This causes an issue when loading a probe and then listing the
tracepoints when no tracing sessions are active: the last probe loaded
is missing from the listing.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>