Prevent re-entrancy in callstack-user context
Userspace callstack context often triggers kernel pagefaults that can be
traced by the kernel tracer which might then attempt to gather the
userspace callstack again... This recursion will be stop by the
RING_BUFFER_MAX_NESTING check but will still pollute the traces with
redundant information.

To prevent this, check if the tracer is already gathering the userspace
callstack and if it's the case don't record it again.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>