statedump: Add a kvm guest statedump
To statedump the guest, the module does the following:

1- Parse the kvm's debugfs directory for entries in the form <pid>-.*
and extract that guest's pid.

2- For each guest pid, read the /proc/<pid>/cmdline file and try to find
a command line argument in the form of -uuid <the guest uuid> and
extract this uuid

3- Trace a kvm guest statedump event with the pid and uuid.

The guest statedump adds ~2us to a kernel with kvm enabled but no
guests. With 1, it is around 15us and 2 guests, around 20us.

Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>