Fix: instruction ptr has different names across arch
Different terms are used to refer to the instruction pointer depending
on the CPU architecture.
For example,
x86 -> ip
powerpc	-> nip
RISC-V -> sepc
ARM -> ARM_pc
Microblaze -> pc

To fix this issue, we use the instruction_pointer() kernel function
(or macro depending on the arch) to get the right field in the pt_regs
struct for the current architecture.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>