statedump: Add thread to interrupt
Threaded IRQs have a 'thread' parameter to the action, defining which
process to wakeup when the IRQ happens.

Having this information will allow to know which process are IRQ
handling process and the analyses can track what happens in those
processes to the IRQ that caused them to wake up.

Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>