statedump: Add thread ID (tid) to interrupt
Threaded IRQs have a 'thread' field set in the the action structure,
defining which process to wakeup when the IRQ happens.

Having this information will allow to know which process are IRQ
handling process and the analyses can track what happens in those
processes to the IRQ that caused them to wake up.

[ Edit by Mathieu Desnoyers: Rename "thread" field to "tid" to stay
  consistent with the rest of lttng-modules tracepoints. ]

Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>