Introduce callstack legacy implementation header
Split the callstack code: keep boilerplate code within the
C implementation file, and move the parts which depend on the
"legacy" (pre-stackwalk) stacktrace kernel API to a separate
implementation header.

This is a preparation step to introduce a new implementation
header for the stackwalk API, added in Linux 5.2 and gradually
integrated within each architecture.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>