Add kernel symbol tracepoints during statedump
This adds an event for each kernel symbol during statedump. It uses the
kallsyms_on_each_symbol kernel function to go through all available
symbols.

Going through the symbols adds a fixed cost to the statedump (9ms on a
machine where statedump usually takes 20ms, so the overhead for the
statedump is slightly lower than 50%) when tracepoints are _not_
enabled.

With the symbol tracepoints enabled, it adds ~135 000 events and more
than doubles the time of the statedump (adds 35ms, so an overhead of
175% compared to statedump without the symbols). It may cause lost
events with the default channel value since the event rate is very dense
during that time.

A trace with only the kernel symbol events has a size of ~5MB on a
normal desktop. In comparison the /proc/kallsyms fils has 6MB.

Signed-off-by: Geneviève Bastien <gbastien@versatic.net>