Add kernel symbol tracepoints when modules come/go
Add tracepoints to enumerate the new symbols brought when a new kernel
module is loaded. It uses the kallsyms_on_each_symbol and filters the
symbols from the new module, as this function is the one exported
from the kernel to list the symbols. It has the same fixed cost from the
function itself (~9ms), whether events are enabled or not.

When the module is unloaded, there is only one event to notify of the
module unload.

Signed-off-by: Geneviève Bastien <gbastien@versatic.net>