Add kernel symbol tracepoints when modules come/go
Add tracepoints to enumerate the new symbols brought when a new kernel
module is loaded. The symbol enumeration is done using the struct
module's kallsyms field. Part of the code for the enumeration is
inspired by or taken from the linux kernel's module.c file. The overhead
when tracepoints are enabled depends on the number of symbols in the
module, but is in tens or hundreds or microseconds. With tracepoints
disabled, overhead is below 10us in the cases tested.

When the module is unloaded, there is only one event to notify of the
module unload.

The lttng_kallsyms_module_coming function is greatly inspired by and a
few lines are copy-pasted from linux's module.c file.

Signed-off-by: Geneviève Bastien <gbastien@versatic.net>