net: Add entry/exit tracepoints for all receive paths
In the linux kernel, there are entry and exit tracepoints on
the various paths of network reception.

Those tracepoints are useful to bound the network reception such that
all events happening between the entry and exit can be linked to this
network reception (for example, wakeups). They can also be used to
perform network stack latency analysis.

Acked-by: Julien Desfossez <jdesfossez@digitalocean.com>
Signed-off-by: Geneviève Bastien <gbastien@versatic.net>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>