Version 2.12.0-pre
Integrate the missing 2.11.0-rc1 entries to the changelog. This
was caused by forking to stable-2.11 the commit before the v2.11.0-rc1
tag. In the future, we will try to keep the rc1 tag as the common
ancestor between master and stable branches.

This also updates the lttng-modules version to 2.12.0-pre, which is
a state that will stay until the v2.12.0-rc1 tag.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>